recording edits

+++ neuron/models/neuron_model_lif_impl.c
@@ -86,7 +86,7 @@ static void lambda(neuron_t * neuron, key_t key, uint32_t neuron_index, bool use
     //nextEventTime = nextEventTime & (~(1 << 31));
     //log_info("Sending Spike with key = %u, neuron_index = %u, payload = %u",key,  neuron_index, nextEventTime );
     log_info("Neuron = %u has fired at time = %u",neuron_index, nextEventTime );
-    neuron->lastThresholdTime = (uint32_t) nextEventTime;
+    neuron->lastThresholdTime = nextEventTime;
     neuron->hasSpiked  = true;
     
     }else if(currentState == 3){
